deleter/pyproject.toml
[project]
name = "deleter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.1.8",
    "faker>=37.1.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]

[project.scripts]
deleter = "deleter.cli.commands:cli"
seeder = "deleter.seeder.commands:cli"
